caf: events: Clean up header files
Clean up sensor_event.h file and move helper
functions, macros and defines to newly introduced
caf_sensor_common.h file.

This is required as a part of NCSDK-17492.
