+++ applications/machine_learning/configuration/nrf52840dk_nrf52840/sensor_manager_def.h
@@ -16,7 +16,7 @@
 const struct {} sensor_manager_def_include_once;
 
 
-static const struct sm_sampled_channel accel_chan[] = {
+static const struct caf_sampled_channel accel_chan[] = {
 	{
 		.chan = SENSOR_CHAN_ACCEL_X,
 		.data_cnt = 1,
